pyproject.toml
[tool.poetry]
name = "PriceTune"
version = "1.0.0"
description = ""
authors = ["TheBreadEnergy <thebreadenergy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-split-settings = "^1.3.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-telegram-bot = "^21.6"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^3.8.0"

[tool.isort]
line_length = 120

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
